Kernel: Resolve the deadlock issue
The cause was, that I was using a KernelMutex for the debug stuff, that
didn't really make sense and was unnecessary.

In the future, I will need to do this properly, however, that will
require careful considerations, about which actions can occur in
interrupt handlers.

However, I immediatelly ran into the next issue with the page allocator.

+++ Kernel/Threads/Scheduler.cpp
@@ -120,7 +120,8 @@ namespace Kernel
 
         m_default_thread->setup_context([&] {
             for (;;) {
-                dbgln("[Scheduler] Running default thread. (refcount={})", m_default_thread->refcount());
+                if (debug_scheduler)
+                    dbgln("[Scheduler] Running default thread. (refcount={})", m_default_thread->refcount());
 
                 MaskedInterruptGuard interrupt_guard;
 
@@ -138,7 +139,8 @@ namespace Kernel
                 // I do not know, if this can happen, better check for it.
                 VERIFY(!dbgln_mutex.is_locked());
 
-                dbgln("[Scheduler] We are about to kill thread '{}' (refcount={})", thread->m_name, thread->refcount());
+                if (debug_scheduler)
+                    dbgln("[Scheduler] We are about to kill thread '{}' (refcount={})", thread->m_name, thread->refcount());
 
                 // Remove the last reference to this thread and thus kill it.
                 VERIFY(thread->refcount() == 1);
